+++ pages/calendar/salary-detail/[id].tsx
@@ -1,8 +1,7 @@
-import { useQuery } from '@tanstack/react-query';
 import { NextPage } from 'next';
 import { useRouter } from 'next/router';
 import ManageDetailScreen from 'src/app.features/calendar/screens/ManageDetailScreen';
-import { getUser } from 'src/app.modules/api/user';
+import useUser from 'src/app.modules/hooks/user/useUser';
 
 interface Props {
 	children: React.ReactElement | null;
@@ -20,22 +19,11 @@ const Admin: React.FC<Props> = ({ children, data }) => {
 const Detail: NextPage = () => {
 	const router = useRouter();
 	const { id } = router.query;
-	const { data, isLoading } = useQuery(['user', 'me'], getUser, {
-		onSuccess: (res) => {
-			// console.log(res);
-		},
-		onError: (error) => {
-			console.log(error);
-		},
-		retry: false,
-		refetchOnMount: false,
-		refetchOnReconnect: false,
-		refetchOnWindowFocus: false,
-	});
+	const { data, isLoading } = useUser();
 	return (
 		<>
 			{!isLoading && (
-				<Admin data={data?.data.data.role}>
+				<Admin data={data.role}>
 					<div className="h-[100vh]">
 						<ManageDetailScreen id={id} />
 					</div>
