+++ src/app.features/calendar/screens/WorkModifyScreen.tsx
@@ -1,19 +1,24 @@
 import React, { useState } from 'react';
 import { useRouter } from 'next/router';
 import SetTimeButtons from 'src/app.components/Button/SetTimeButtons';
-import { useMutation } from '@tanstack/react-query';
 import Modal from 'src/app.components/Modal/Modal';
 import Overlay from 'src/app.components/Modal/Overlay';
 import useModalStore from 'src/app.modules/store/modal';
 import Header from 'src/app.components/Header';
 import DelIcon from 'src/app.modules/assets/calendar/delete.svg';
 import Bar from 'src/app.components/app.base/Button/Bar';
+import { crossDate } from 'src/app.modules/util/calendar';
+import { MutateTpye } from 'src/app.modules/api/client';
 import useStore from '../store';
 import useTimeSetStore from '../store/time';
-import { delWorkModify, postWork, putWorkModify } from '../api';
+import { delWorkModify, MutateBody } from '../api';
 
 type Flag = 'startTime' | 'endTime' | null;
-function WorkModifyScreen() {
+interface Props {
+	WorkMutate: MutateTpye<MutateBody>;
+	ModifyMutate: MutateTpye<MutateBody>;
+}
+function WorkModifyScreen({ WorkMutate, ModifyMutate }: Props) {
 	const { isModalOpen, modalIsOpen } = useModalStore();
 	const [workTime, setWorkTime] = useState<string>('');
 	const { clickDay, toDay, workDay, isDayReset } = useStore();
@@ -23,12 +28,12 @@ function WorkModifyScreen() {
 		user: { startTime, endTime },
 		setTime,
 	} = useTimeSetStore();
+	const { clickDayData, toDayData } = crossDate(clickDay, toDay);
 	const [openModalFlag, setOpenModalFlag] = useState<Flag>(null);
 	const timeHandler = (e: React.BaseSyntheticEvent) => {
 		const {
 			target: { name, value },
 		} = e;
-
 		setTime(value, name, openModalFlag as 'startTime' | 'endTime');
 	};
 
@@ -47,24 +52,6 @@ function WorkModifyScreen() {
 		}
 	};
 
-	// 출근하기
-	const { mutate: WorkMutate } = useMutation(postWork, {
-		onSuccess: (res) => {
-			console.log(res);
-			router.back();
-		},
-		onError: (error) => alert('오류 발생.'),
-	});
-
-	// 수정하기
-	const { mutate } = useMutation(putWorkModify, {
-		onSuccess: (res) => {
-			// console.log(res);
-			router.back();
-		},
-		onError: (error) => console.log(error),
-	});
-
 	// 수정 버튼
 	const modifyBtn = () => {
 		const workTimeData = getWorkTimeString();
@@ -72,13 +59,13 @@ function WorkModifyScreen() {
 		const startSplit = Number(start.split(':')[0]) * 60 + Number(start.split(':')[1]);
 		const endSplit = Number(end.split(':')[0]) * 60 + Number(end.split(':')[1]);
 		const timeDiff = Math.abs((startSplit - endSplit) / 60);
-		if (!workDay && new Date(clickDay) < new Date(toDay)) {
+		if (!workDay && clickDayData < toDayData) {
 			// 출근하기
 			WorkMutate({ year, month, day, workTime: workTimeData, workHour: timeDiff });
 		} else {
-			mutate({ year, month, day, workTime: workTimeData, workHour: timeDiff });
+			// 수정하기
+			ModifyMutate({ year, month, day, workTime: workTimeData, workHour: timeDiff });
 		}
-
 		isDayReset();
 	};
 
@@ -159,11 +146,9 @@ function WorkModifyScreen() {
 				</div>
 			</div>
 			{isModalOpen && (
-				<>
-					<Overlay>
-						<Modal content="출근기록이 삭제됩니다!" deleteFn={() => delBtn()} />
-					</Overlay>
-				</>
+				<Overlay>
+					<Modal content="출근기록이 삭제됩니다!" deleteFn={() => delBtn()} />
+				</Overlay>
 			)}
 		</>
 	);
