+++ src/app.modules/util/calendar.ts
@@ -13,3 +13,11 @@ export function getDaysInMonth(year: number, month: number) {
 	const day = new Date(year, month + 1, 0).getDate();
 	return `${month + 1}.1 - ${month + 1}.${day}`;
 }
+
+export function crossDate(clickDay: string, toDay: string) {
+	const [currentYear, currentMonth, currentDay] = clickDay.split('.').map(Number);
+	const [toDayYear, toDayMonth, toDayDay] = toDay.split('.').map(Number);
+	const clickDayData = new Date(currentYear, currentMonth, currentDay, 0, 0, 0);
+	const toDayData = new Date(toDayYear, toDayMonth, toDayDay, 0, 0, 0);
+	return { clickDayData, toDayData };
+}
